+++ walk-in-interview/src/main/webapp/apis.js
@@ -10,5 +10,5 @@ export const API = {
   'expire-job': '/jobs/expire',
   'business-log-in': '/business-log-in',
   'applicant-log-in': '/applicant-log-in',
-  'show-job-posts-made': '/my-jobs',
+  'business-jobs-list': '/my-jobs',
 };
