+++ walk-in-interview/src/main/webapp/business-jobs-list/index.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta charset="UTF-8">
-    <link rel="stylesheet" href="./style.css">
+    <link rel="stylesheet" href="../style.css">
   </head>
 
   <body>
@@ -33,5 +33,5 @@
     </div>
   </body>
 
-  <script type="module" src="./show-job-posts-made.js"></script>
+  <script type="module" src="./script.js"></script>
 </html>
